move subpage related files into /SubPage and pull gene text formatting into EditingDesignSubPage
